refactor(redis): 将 DBSize 数据类型从 Long 改为 Int
- 修改了 DBNode 类中的 dbSize 字段类型从 long到 int
- 更新了 JedisClusterClient 和 JedisPoolClient 中 dbSize 方法的返回类型从 Map<Integer, Long> 到 Map<Integer, Integer>
- 调整了 RedisClient 接口中 dbSize 方法的返回类型
- 更新

+++ rdm-redis-imp/src/main/java/xyz/hashdog/rdm/redis/imp/Util.java
@@ -207,10 +207,10 @@ public class Util extends xyz.hashdog.rdm.common.util.Util {
      * @param row keyspace
      * @return db,size
      */
-    public static Tuple2<Integer, Long> keyspaceParseDb(String row) {
+    public static Tuple2<Integer, Integer> keyspaceParseDb(String row) {
         String[] a = row.split(":");
         int db =Integer.parseInt(a[0].substring(2));
-        long size =Long.parseLong(a[1].split(",")[0].substring(5)) ;
+        int size =Integer.parseInt(a[1].split(",")[0].substring(5)) ;
         return new Tuple2<>(db,size);
     }
 
@@ -220,7 +220,7 @@ public class Util extends xyz.hashdog.rdm.common.util.Util {
      * @param value keyspace
      * @return db,size
      */
-    public static Tuple2<Integer, Long> keyspaceParseDb(String key, String value) {
+    public static Tuple2<Integer, Integer> keyspaceParseDb(String key, String value) {
         return keyspaceParseDb(key+":"+value);
     }
